requirements.txt
# Core dependencies
python-dotenv>=1.0.0
kivy>=2.1.0
flask>=3.0.0
requests>=2.28.0
Pillow>=10.0.0
reportlab>=4.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
freezegun>=1.2.0

# Development dependencies (optional but recommended)
black>=23.0.0
flake8>=6.0.0
